Application processes

+++ DGEFramework/Processes/ApplicationProcess.cs
@@ -33,6 +33,8 @@ namespace DGE.Processes
                 this.safeStopMethod = safeStopMethod;
 
             ApplicationManager.Add(this);
+
+            logger = new EnderEngine.Logger($"Process:{Id}");
         }
 
         public override void Dispose()
@@ -42,8 +44,12 @@ namespace DGE.Processes
 
         public override void Start()
         {
+            if (!ProcessExitedCleanUp() || status >= ApplicationStatus.ON)
+                return;
+
             process = new Process { StartInfo = startInfo };
             process.EnableRaisingEvents = true;
+            process.Exited += (s, e) => ProcessExitedCleanUp();
             process.Exited += OnStopped;
 
             OnStarting?.Invoke(this, EventArgs.Empty);
@@ -64,15 +70,45 @@ namespace DGE.Processes
         {
             OnShutdown?.Invoke(this, EventArgs.Empty);
 
-            if (safeStopMethod is null)
-                process?.Kill();
-            else
-                safeStopMethod(this);
+            if (ProcessExitedCleanUp() || status <= ApplicationStatus.STOPPING) // If the process quit already
+                return;
 
-            process?.Dispose();
-            process = null;
+            try
+            {
+                if (safeStopMethod is null)
+                    process?.Kill();
+                else
+                {
+                    logger.Log($"Invoking existing process SafeStop method", EnderEngine.Logger.LogLevel.INFO);
+                    safeStopMethod(this);
+                }
+            }
+            catch (Exception e)
+            {
+                logger.Log($"Stop failed (Is the process exited already ?) : {e.Message}", EnderEngine.Logger.LogLevel.ERROR);
+            }
+
+            ProcessExitedCleanUp();
 
             OnStopped?.Invoke(this, EventArgs.Empty);
         }
+
+        /// <summary>
+        /// Cleans up pocess memory usage if it has exited
+        /// </summary>
+        /// <returns> Returns if it could clean up memory / or it is already cleaned up, false if it couldn't (could mean the process is still running) </returns>
+        private bool ProcessExitedCleanUp() //This function is used redondently to prevent bugs
+        {
+            if (process is null)
+                return true;
+            if (process.HasExited)
+            {
+                process?.Dispose();
+                process = null;
+                return true;
+            }
+            return false;
+        }
+
     }
 }
